feat: implement scroll progress fallback and refactor ScrollProgress component
