+++ src/components/App/ScrollProgress/Fallback.tsx
@@ -1,0 +1,13 @@
+'use client';
+
+import { useScrollProgressFallback } from './useScrollProgressFallback';
+
+type Props = {
+  targetId: string;
+};
+
+export function ScrollProgressFallback({ targetId }: Props) {
+  useScrollProgressFallback(targetId);
+
+  return null;
+}
